+++ External/Plugins/FlashLogViewer/PluginUI.cs
@@ -158,7 +158,7 @@ namespace FlashLogViewer
             this.logTextBox.BorderStyle = System.Windows.Forms.BorderStyle.None;
             this.logTextBox.DetectUrls = false;
             this.logTextBox.Dock = System.Windows.Forms.DockStyle.Fill;
-            this.logTextBox.Font = new System.Drawing.Font("Courier New", 8.75F, System.Drawing.FontStyle.Regular, System.Drawing.GraphicsUnit.Point, ((byte)(0)));
+            this.logTextBox.Font = new System.Drawing.Font("Courier New", 8.75F, System.Drawing.FontStyle.Regular, System.Drawing.GraphicsUnit.Point, 0);
             this.logTextBox.Location = new System.Drawing.Point(1, 30);
             this.logTextBox.Name = "logTextBox";
             this.logTextBox.ReadOnly = true;
