+++ PluginCore/SevenZip/Compress/RangeCoder/RangeCoder.cs
@@ -64,7 +64,7 @@ namespace SevenZip.Compression.RangeCoder
 
         public void ShiftLow()
         {
-            if ((uint)Low < (uint)0xFF000000 || (uint)(Low >> 32) == 1)
+            if ((uint)Low < 0xFF000000 || (uint)(Low >> 32) == 1)
             {
                 byte temp = _cache;
                 do
